separated out RedirectableReceiver into its own class file
